Change the message in UI for NETWORK_NOT_FOUND event
NETWORK_NOT_FOUND event is sent for networks failing to
connect due to, failed to find a network in supplicant scan result.
This can happen due to two reasons,
1. Network is seen in supplicant scan result. But the network
capabilities doesn't match with the network configuration sent by
framework.
2. During connect, if supplicant scan cache is not recent enough
(<5 seconds), it issues a scan. If the AP is far from device, this
scan may fails to find the network.

"Network not compatible with this device" message in UI will confuse
the user especially when the network is not seen due to weak signal.
So changed the message in UI to a generic string
"Can't connect, try again later"

Bug: 191061519
Test: Manual - Tried connecting to a reconfigured AP and verified that
the "Can't connect. Try again later" message is displayed.

Change-Id: Ief7aafb47346599eabf8076f3c73c78e6aa48fa8