WifiTrackerLib: Populate network request scans after initiation
NetworkRequestEntry scans should be populated immediately after
initiation. This is because the NetworkRequestEntry is created once
WifiManager.getConnectionInfo() indicates a connection to a network
request. Since the connected WifiInfo will have invalid RSSI for a short
period and we might not trigger a scan update for another several
seconds, we should populate the entry with the scans we already have in
order to display a non-empty RSSI.

Bug: 195986350
Test: manual shell command to add network request for glob SSID with
wrong password, tap on network in network request dialog, and verify
RSSI remains non-empty.

Change-Id: I083a18ef7305f161a121634fd0f0f4833a417783