WifiTracker: Update isDefaultNetwork for MBB roaming
When a new wifi network is connected to, we mark the entry as not
default and only mark it as default when the default network callback
changes to wifi. However, in MBB we may always be connected to wifi as
the default route and never get the default network callback after the
roam. Thus we should always update the newly connected entry with the
default network status whenever we change the connected entry.

Bug: 199809298
Test: atest WifiPickerTrackerTest
Change-Id: Id4247133be6ba6d1f1514e8f3d24333831201bca
(cherry picked from commit 370c0f6314bd83f73b7561f75b682525426b2a59)
Merged-In: Id4247133be6ba6d1f1514e8f3d24333831201bca