Use specific comparators for WifiEntry sorting
Remove the default WifiEntry.compareTo() and introduce specific
comparators to sort WifiEntries for either WifiPickerTracker or
SavedNetworkTracker.

Comparator for WifiPickerTracker is inherited from
WifiEntry.compareTo(), with the addition that networks that
cannot be connected to (such as due to enterprise admin
restrictions) should be sorted to the bottom of the picker.

Comparator for SavedNetworkTracker simply sorts by Title and does not
sort based on level or whether the network is connected or connectable.

Bug: 216859324
Test: atest WifiTrackerLibTests SdkWifiTrackerLibTests
Change-Id: I230383a1af614807da4fd58163e1a9b41cd5b4c8