Filter out Passpoint security type from StandardWifiEntry
Starting in T, ScanResult.getSecurityTypes() is used to get the list of
supported security types of a scan result. This now includes
SECURITY_TYPE_PASSPOINT_R1_R2/R3, which results in a StandardWifiEntry
with invalid security type, resulting in UI issues. Thus these
passpoint security types should be ignored, and the scan should be
treated as a normal EAP network.

Bug: 219106254
Test: atest UtilsTest, manually verify EAP/Passpoint scan result
displays correctly as an EAP network.

Change-Id: I6a5928aa0256aef99d6a0a74fe1202e0437bdde4