[WTL] Register callbacks on worker thread
Registering the broadcast receiver and network callbacks in the main
thread causes race conditions if the receiver/callbacks are called
before onStart()->handleOnStart() is called. This may cause an invalid
initial state for WifiEntries, such as not being the default route even
when it is.

Instead, we should register these callbacks in the worker thread, to
ensure that onStart()->handleOnStart() is always called before any of
the callback methods are run, which simplifies the threading logic.

Bug: 222527907
Test: atest WifiTrackerLibTests SdkWifiTrackerLibTests, manually verify
opening Network Details page of default route connected network does not
show "!".

Change-Id: I7564710767e90bf4218d03ab1c4cb0cd9e9ab908