WifiTrackerLib: Always return true for isPrimary for SUW
SUW cannot access isPrimary since it doesn't hold the necessary
NETWORK_SETTINGS permission. However, allowing SUW to access this
requires a new REDACT_FOR_SETUP_WIZARD flag, which cannot be added in T
since it would be a new API.

Thus we should always return true for isPrimary for SUW to unblock SUW's
WifiTrackerLib migration for T. Note that this will be the same behavior
as WifiTracker1, which does not distinguish between primary and
secondary network, so there is no regression from SUW's perspective.

Bug: 227534637
Test: atest WifiTrackerLibTests SdkWifiTrackerLibTests
Change-Id: I8bdef4db1bd4bbcb6c798b690888bc4bbb34ef0a