Show "Checking for internet access..." if network not validated yet
Wifi picker currently shows "Connected to device. Can't provide
internet" for any network that doesn't have validation. However, this is
confusing for networks that are expected to have internet but are still
in the middle of validation. Thus we should show "Checking for internet
access..." while we're in the middle of validation, and then show the
other message once the validation attempt fails.

Bug: 230794429
Test: atest WifiTrackerLibTests, manually verify "Checking for internet
access..." appears in Wifi picker network summary.

Change-Id: I00d8061eaa75123ed1b8e6edf118363c74d8fde7