Show password dialog for authentication issues if never connected
Wrong credentials may cause DISABLED_AUTHENTICATION_FAILURE, which
doesn't prompt a wrong password dialog until 3 autoconnect failures.
Instead of waiting for 3 failures, we should infer that the credentials
are wrong and show the dialog if we have any number of
DISABLED_AUTHENTICATION_FAILURE and we've never connected before.

Bug: 271535438
Test: atest StandardWifiEntryTest, manually verify connection to EAP
network with wrong credentials will show a dialog the next time we try
to connect.

Change-Id: Iba5859ab6da4d5af64bb1cba565e196e84c670ec