Fix Passpoint entry showing without lock
Passpoint entries are showing in the wifi picker without a lock due to
mistakenly returning SECURITY_NONE due to
Utils#getSingleSecurityTypeFromMultipleSecurityTypes(...) not handling
multiple passpoint security types. Fix this method by always defaulting
to the first security type in the list if we don't have special handling
for it them.

Bug: 273766638
Test: atest UtilsTest, manually verify passpoint network shows with lock
Change-Id: Icf3f6a3082fb19b54b9cd6eb3469a31282f78183