Wrap mActiveWifiEntries in a synchronized block for the linter
mActiveWifiEntries is protected by mLock and should only be accessed
while the lock is acquired. However, the linter complains about access
to mActiveWifiEntries in a stream filter in updateWifiEntries(), even
though the stream is only used inside a synchronized block. Add another
synchronized block in the stream filter to make the linter happy.

Bug: 270173046
Test: atest WifiPickerTrackerTests
Change-Id: Ib3339089511d64c01f8e5894ed1c667634187a4e