Cut off scan loop if we're already stopped
BaseWifiTracker.Scanner is stopped on onStop(), but will start up a new
scan loop if Scanner.start() is called again via a
WIFI_STATE_CHANGED broadcast that was queued before the broadcast
receiver was unregistered. To avoid infinitely scanning from this, have loop
stop itself if we detect that we've already called onStop().

Bug: 274017616
Test: atest WifiPickerTrackerTest
Change-Id: I7d6c9d06b10b1bbe1f125e83bb9287886955a80d