Use separate LifecycleObserver class in WifiTrackerLib
LifecycleObserver requires that all of its declared methods only use
APIs available in the current SDK version. Since BaseWifiTracker
declares some methods using APIs that might not be available on the
current platform (i.e. SUW using the latest WifiTrackerLib on an S
device), we need to move the LifecycleObserver to a separate class.

Bug: 278666253
Test: atest WifiTrackerLibTests
Change-Id: Ib20845b9fb8059e1d1eadb77c75bd66566d56f0f