Reformat scan result description for MLO
Currently scan result description displays all details of MloLink for
each of the affiliated links. Some of the details are not relevant
during scanning stage. Also some information is repeated as each of the
link AP has it's own scan entries.

Current scan description is,
 bssid1=freq1,RSSI,Wi-Fi 7,mldMac,linkId,affLinks[MloLink{Band,channeld,
	id,state,rssi,link speed...},MloLink{..} ..]
 bssid2=freq2,RSSI,Wi-Fi 7,mldMac,linkId,affLinks[MloLink{Band,channeld,
	id,state,rssi,link speed ...},MloLink{..} ..]
 bssid3=freq2,RSSI,Wi-Fi 7,mldMac,linkId,affLinks[MloLink{Band,channeld,
	id,state,rssi,link speed...},MloLink{..} ..]

The MLO state, link speed ..etc. is not relevant during scan stage.
Other information like link RSSI, link ID ..etc. are repeated in the AP
scan details.

Reformat the description to make it short and more readable.
 bssid1=freq,RSSI,Wi-Fi 7,mldMac,linkId,affLinks[{bssid1,freq1},
 		{bssid2,freq2},{bssid3,freq3}]
 bssid2=freq,RSSI,Wi-Fi 7,mldMac,linkId,affLinks[{bssid1,freq1},
 		{bssid2,freq2},{bssid3,freq3}]
 bssid3=freq,RSSI,Wi-Fi 7,mldMac,linkId,affLinks[{bssid1,freq1},
 		{bssid2,freq2},{bssid3,freq3}]

Though the MAC address and frequency of each of the affiliated links are
repeated, it is still useful in case some of the link AP's are missed in scan.

Bug: 273355708
Test: m
Change-Id: I46cae60527117ddc42d94716832dd80d7e88b6ca