Go back to using app default network instead of the system default
Go back to using registerDefaultNetworkCallback instead of
registerSystemDefaultNetworkCallback so we can properly detect if VPN
over wifi is the default network, and show W+ as connected.

However, going back to registerDefaultNetworkCallback will result in
b/224857908 and show wifi as low quality if it's validated but the VPN
app chooses cell as the underlying network. To prevent this, do not show
wifi as low quality if cell is default via VPN.

Bug: 265362819
Test: atest WifiTrackerLibTests
Change-Id: Id3f840ac28ae37afabf930efeec013b391e23030