Add getLifecycleObserver() to allow registration after construction
BaseWifiTracker registers a LifecycleObserver within its constructor,
but if the LifecycleObserver.onStart/onStop/onDestroyed is called before
the constructor is complete, then we get an NPE trying to reference
BaseWifiTracker.this. To avoid this case, allow passing a null Lifecycle
in the constructor and registering the LifecycleObserver from
getLifecycleObserver() afterwards.

Bug: 287876602
Test: atest WifiTrackerLibTests
Change-Id: I0cfb1789d1f2e23b7b38ce676f501b25e1e2cc43