Report full scan result for WifiScanner request
The WifiScanner request should include flags for reporting the scans as
a full scan result. This is to prevent the case where wifi is enabled,
causing Settings to issue a scan request before the connectivity scan,
causing a delay in auto-connect since the Settings scan isn't reported
as full results.

Bug: 290719143
Test: build
Change-Id: I28c003c20e2fc7df04c126f5ab675cad536c0cbc