Remove timer-based connect failure callback
When tapping on a network, we check if we're connected to the network
10 seconds after WifiManager.connect() succeeds, and notify the
ConnectCallback of a failure if we aren't connected at that point.
However, if the user taps on a different network before the connection
is over, then we detect that we aren't connected to the original network
and notify ConnectCallback of a failure, resulting in a "Failed to
connect to network" toast.

To prevent this confusing toast, we should remove the timer-based system
and only notify failure if WifiManager.connect() fails. The toast is
largely redundant since
1) The user may miss the toast
2) We already show a notification for wrong password
3) We already display failure reasons in the network summary

Bug: 291974248
Test: build, manually verify "Failed to connect to network" toast
doesn't appear when tapping on different networks before a manual
connection is complete.

Change-Id: Ie93365efeb79e56b0f495e9f60aaf35fcfab782c