Override getLevel for HotspotNetworkEntry.
Level is set to WIFI_LEVEL_UNREACHABLE by default which
causes WiFi slice to filter it out.
Changed getLevel in HotspotNetworkEntry to return max value
when not connected so it will not be filtered and still show
level correctly when connected.

Bug: 268550769
Test: atest com.android.wifitrackerlib.HotspotNetworkEntryTest

Change-Id: If81cfee11dcd6bbc8dd0779b77ddb01ba0977f62
Merged-In: Ied0efea246b555144f0f04bb4743c234c0163b7d