Show "Available via..." for captive portals even if Wi-Fi is default
Captive portal indicates that the wifi network isn't providing internet
access, even if wifi is the default network. Thus we should show
"Available via app / Sign in to network" instead of "Connected via
app / Sign in to network" for suggestion networks with captive portals
with wifi as the default network.

Bug: 293946849
Test: atest WifiTrackerLibTests
Change-Id: I1ed148a33e98e6c230808c74008d7a39f4568fc5