Enable WifiTrackerLib verbose logging for userdebug builds
Enable WifiTrackerLib verbose logging for userdebug builds, but keep
 the verbose summary tied to the wifi service verbose logging level.

We also need to change the verbose logging flag from static to
non-static, since the static BaseWifiTracker object is shared between
all clients in SystemUI, including the QuickSettings which we do want to
show verbose logging for.

Bug: 301199999
Bug: 279512950
Test: atest WifiTrackerLibTests, manually verify verbose logging (but
not summaries) is shown for userdebug build when wifi verbose logging is
disabled.

Change-Id: I3464d56c704e4b5dbb386f26c87283d8b73bd279