Add WifiEntry.needsWifiConfiguration()
Add WifiEntry.needsWifiConfiguration() to indicate if a network needs to
have a WifiConfiguration saved before connecting. This is to provide a
synchronous way to query if we need a config instead of waiting for
CONNECT_STATUS_FAILURE_NO_CONFIG, in case the caller (e.g. Android TV)
wants to show a progress spinner upon a user press.

Bug: 305092860
Bug: 153627947
Test: atest StandardWifiEntryTest
Change-Id: I9065fe75c1b21a5b0e2b9045e0d0635f551d8c81