Restore connected state in Wifi picker.
This part of the change reads the connected state from the extras bundle and updates the HotspotNetworkEntry. The entry uses this state to reset the mCalledConnect boolean and trigger the callback.

The connected enum value is defined locally in WifiTrackerLib, but will be moved to the SharedConnectivity API on the next API update.

BUG: 312561529
Change-Id: I6675024c67f926d28544013b23cef88b2d63ce6e