Fix mLastNetwork being cleared upon second new network callback
During network roaming, the wifi network may change but the default
network may remain on the old network for a few seconds. To avoid
flickering the connected network to "not default", we keep the pre-roam
network in WifiEntry.mLastNetwork to match with the old default network.
Fix an issue where mLastNetwork is cleared if the network capabilities
change but the network remains the same.

Flag: EXEMPT minor UI fix

Bug: 405915182
Test: atest StandardWifiEntryTest
Change-Id: I4202deaf0e9fcb19d8a8bbb9ad1d2f8395189712