Use old scans when going from Internet picker -> Network Details
Internet picker will show scan results up to 5 minutes old when faced
with consecutive scan failures, but prune scans older than 15 seconds
if a scan is successful. If a user taps on a network while the scans are
failed, then the transition to the NetworkDetailsTracker will prune the
latest scans to 15 seconds, causing the network to briefly appear out of
range.

Fix this by avoiding pruning the scans when starting a new
NetworkDetailsTracker.

Flag: EXEMPT minor UI fix

Bug: 394615022
Test: manually verify tapping on network after several scan failures
will not show it as disconnected in the Network Details page.

Change-Id: Ie2c8b380be34f169f0c1f45a325a14032566382a