Don't show TDI error message if a connection is available
When a WPA3-Transition or OWE-Transition AP sends out a TDI, the weaker
security type (WPA2 or Open) is disabled until the user removes the
network, and a "Unavailable with current security settings" message is
shown. However, this message is confusing if there is a connectable WPA3
or OWE AP in range, so avoid showing this message in this case.

Flag: EXEMPT minor UI change

Bug: 419014346
Test: atest StandardWifiEntryTest
Change-Id: I18d0b49bd5026255da45ac73a1d49b94cd93c617