Remove stale connected WifiEntry in handleOnStart
A connected WifiEntry that disconnects during screen off may appear in
the picker list after screen on for a few seconds even if there are no
recent scan results for it. This is due to the stale WifiInfo preventing
the scan update logic from removing it as an out-of-range network.

Fix this by making sure the WifiInfo is cleared before updating the scan
info.

Flag: EXEMPT minor UI fix

Bug: 427107455
Test: atest WifiPickerTrackerTest
Change-Id: I035d9e9b6ffeeba15f0aef45fb75fc26c7051bce