Make thread-shared booleans volatile
BaseWifiTracker relies on some booleans that are set by the main thread
to be read by the worker thread in order to gate operations such as
scanning for wifi. These should be marked "volatile" so that the values
area always flushed from the CPU cache to main memory, ensuring that the
worker thread will read the latest value.

Flag: EXEMPT minor bugfix

Bug: 435135241
Test: atest WifiTrackerLibTests
Change-Id: Ie3712b94eaf26c5c618f6db5c6cf92fdef563544