Wrap startScan in try-catch for user-switching case
User switch may happen before the scanning mechanism stops, which will
cause a SecurityException due to background user requesting the scan.
Although we already have the shouldScan() flag check to prevent this,
wrapping the actual call in a try-catch will add an extra layer of
protection in case the system state is out of sync.

Flag: EXEMPT minor bugfix

Bug: 435135241
Test: atest WifiTrackerLibTests
Change-Id: I45dc98a514c29c1293fe7594684971969c6e6413