Don't restrict wifi for carrier-merged if wifi is off
Tapping on the SIM network in the Settings page will call into
MergedCarrierEntry.connect(), which starts restricting autojoin for
non-carrier networks (which is intended to signal the user's preference
to stay connected to the carrier). However, this action doesn't make
sense when Wi-fi is off since the user is already connected to the
carrier, and it also shows a misleading "Wi-Fi won't autoconnect" toast.

Instead, we should avoid setting this restriction and showing the toast
if Wi-Fi is off. Then when the user turns on Wi-Fi again, they're free
to autoconnect to their saved Wi-Fi networks.

Flag: EXEMPT minor bugfix

Bug: 438831994
Test: atest MergedCarrierEntryTest
Change-Id: I218a28636ec6dd6a373858deabc497aab55c67c0