Fix callback registration race
BaseWifiTracker registers its network callbacks and broadcast receiver
on the worker thread since ag/17085223 in order to make sure none of the
sticky broadcasts/callbacks were called before handleOnStart. However,
this results in unreliable callback unregistration due to races between
the worker thread and the onDestroy method which unconditionally
unregisters everything on the main thread.

Fix this by going back to handling registration/unregistration on the
main thread, while making sure that handleOnStart is posted before
registration to ensure that it is called before any of the
callbacks/broadcasts.

Flag: EXEMPT minor bugfix

Bug: 436462749
Test: atest WifiTrackerLibTests
Change-Id: I74450bf5c72334519c334bf95119ba12a930c218