Fix MergedCarrierEntry initialization race
MergedCarrierEntry may be initialized either on the worker thread or on
demand from the main thread if getMergedCarrierEntry() is called before
the worker thread initialization. However, this causes a race condition
if both threads initialize at the same time, causing MergedCarrierEntry
to sometimes be left in a stale state. Fix this by synchronizing both
initializations and making mMergedCarrierEntry volatile.

Flag: EXEMPT minor bugfix

Bug: 429028697
Test: atest WifiTrackerLibTests
Change-Id: Ie9706ca278a373f4086366829f8bd2f759da0eff