Remove app label fallback for carrier network summary
Currently we fallback to "Connected via <App name>" if the carrier name
of a WifiConfiguration is not found, but this results in misleading UI
(e.g. "Connected via Google Wi-Fi Provisioner" instead of the actual
carrier name).

Fix this by removing this fallback and simply showing "Connected" if the
carrier name cannot be found. Since it is unexpected that the carrier
name is not found, also add error logs for debugging such cases.

Flag: EXEMPT BUGFIX

Bug: 433296627
Test: atest WifiTrackerLibTests
Change-Id: Ia24abf108df57f5b1e11d12e516349d05396f188