Show "Connected via" for non-default Wifi networks
Internet Settings page generally reserves the string "Connected" for the
network that is providing the default internet route. However, this
leads to misleading UI for suggestion/specifier networks, which the user
might want to explicitly check the connection state of even if internet
is not provided (such as for the Android Auto case). Thus, always show
"Connected via <suggestion/specifier app>" for those networks even if
internet is not being provided. Note that any network errors that result
in no interent will already be conveyed to the user via additional
strings (e.g. "No internet access").

Flag: EXEMPT BUGFIX

Bug: 440191301
Test: atest UtilsTest
Change-Id: Id2da48d3e959653a8e595abc552eceecdc9e432c